M18.6 (4/5) + M19.5.2 polish + close pending M20.5 code
M18.6 — SMP polish, 4 of 5 sub-items shipped (.5 MSI/MSI-X remains):

  .1 Per-CPU runqueue + load balancer.  Global runqueue replaced by an
     intrusive doubly-linked per-CPU rq rooted at percpu->rq_head and
     threaded via task->rq_next/rq_prev.  Master task list (for ps /
     iteration / find) split out, threaded via task->next under its
     own master_lock.  Load balancer runs from schedule's idle-fallback
     path: when local rq is empty, scan peers, steal a task whose
     affinity allows running here.  Scheduler lock acquire/release is
     correct under task migration: schedule's unlock re-reads
     this_cpu()->rq_lock so the lock released is whichever CPU we are
     on NOW (Linux finish_task_switch pattern).
  .2 Per-CPU preempt_count.  Was a single global, which was
     SMP-incorrect (disabling on CPU A also gated CPU B).  Now lives
     in struct percpu; accessors bracket the RMW in hal_intr_save /
     restore so the local timer cannot migrate us mid-increment.
  .3 Task affinity + taskset.  cpu_mask per task (default any-CPU);
     scheduler and steal-time both filter by mask & (1<<this_cpu_id).
     task_set_affinity migrates if current cpu_home is no longer in
     the new mask.  Shell: taskset <pid> <hex-mask>.  lscpu now also
     prints per-CPU rq depth.
  .4 Cross-CPU preempt IPI sender.  New lapic_send_ipi (fixed-delivery,
     self-IPI no-op-ed internally) + smp_send_reschedule wrapper on
     vector 0x41 (handler already wired since M18.5).  task_enqueue
     fires it when a task lands on a CPU other than self: wakes the
     target hlt-ed idle so the task starts running without waiting up
     to 10 ms for the next local LAPIC tick.

M19.5 — Memory polish, 1 of 3 sub-items shipped (.1 HIGHMEM and .3
SRAT/NUMA remain):

  .2 Empty-slab caching.  Each slab cache keeps up to EMPTY_SLAB_MAX
     (=4) fully empty slabs in a per-cache LIFO instead of immediately
     releasing to the buddy.  Refill prefers a cached empty slab over
     a fresh page_alloc.  slabinfo gained a CACHED-EMPTY column.

Closes pending M20.5 Phase A/B/C code that was authored but never
committed: kernel/hal/x86_64/{ap_trampoline.s, smp.c, syscall.c,
usermode.s} (new files), the kernel/core/syscall.c -> kernel/hal/x86/
syscall.c rename, x86_64 vmm.c US-bit fix, kprintf %l/%ll/%z/%z, the
LAPIC + IOAPIC arch-family-share move, and the Makefile entries.

Verified end-to-end on i386 -smp 4 and x86_64 -smp 4: preempt
self-test PASSes (hog runs while kernel_main halts); parallel
self-test PASSes (both hogs make progress on N-CPU systems).

Pitfalls codified in source / DOCS.md change log:
* schedule()'s unlock must re-read this_cpu() — context_switch may
  have resumed us on a different CPU than the one we entered on.
* IPI on vector 0x41 must NOT set need_resched on the receiver — its
  purpose is to wake hlt; the receiver's idle loop then yields
  naturally.
* slab_release re-increments c->slabs when keeping a page cached;
  caller pre-decrements before calling.

Deferred to a follow-up polish session: §M18.6.5 (MSI/MSI-X),
§M19.5.1 (HIGHMEM + kmap), §M19.5.3 (SRAT/NUMA), §M20.6.1
(SYSCALL/SYSRET — needs GDT slot reorg), §M20.6.2/3 (xHCI + virtio-
blk 64-bit DMA audit + x86_64 enable).  All independent of M21+.

+++ kernel/hal/x86/ioapic.c
@@ -63,12 +63,12 @@ static void io_w(uint32_t reg, uint32_t v) {
     *(volatile uint32_t*)(g_ioapic_mmio + IOAPIC_OFF_IOWIN) = v;
 }
 
-int ioapic_init(uint32_t phys, uint32_t gsi_base) {
-    uint32_t aligned = phys & ~0xFFFu;
+int ioapic_init(uintptr_t phys, uint32_t gsi_base) {
+    uintptr_t aligned = phys & ~((uintptr_t)0xFFFu);
     if (vmm_map(aligned, aligned, VMM_WRITABLE | VMM_CACHE_DIS) != 0) {
         /* Probably already mapped — proceed. */
     }
-    g_ioapic_mmio = (volatile uint8_t*)(uintptr_t)phys;
+    g_ioapic_mmio = (volatile uint8_t*)phys;
     g_gsi_base    = gsi_base;
 
     uint32_t ver = io_r(IOAPIC_REG_VERSION);
